Cleanup all unnecessary assets for archive sites
This will build a minimal site for docs (and add-on
reference) which will redirect back to the main site
for all other top-level navigation.

Set the OH_DOC_VERSION environment variable, run
prepare-docs.rb manually then `vuepress build`

The resulting website is under `vuepress`.
Currently it is to be hosted on the main website
under `.vuepress/public/v2.x`.

Signed-off-by: Yannick Schaus <github@schaus.net>